feat(209-03): narrow day-index uint48 to uint32 in DegenerusJackpots, DegenerusVault, DeityBoonViewer
- DegenerusJackpots: lastBafResolvedDay storage variable
- DegenerusVault: openLootBox interface param, gameOpenLootBox param
- DeityBoonViewer: deityBoonData return type, deityBoonSlots return + local var
- Full forge build passes (only pre-existing FuturepoolSkim.t.sol failure)

+++ contracts/DegenerusJackpots.sol
@@ -128,7 +128,7 @@ contract DegenerusJackpots is IDegenerusJackpots {
     mapping(uint24 => mapping(address => uint256)) internal bafPlayerEpoch;
 
     /// @notice Day index of the most recent BAF jackpot resolution (any bracket).
-    uint48 internal lastBafResolvedDay;
+    uint32 internal lastBafResolvedDay;
 
     /*+======================================================================+
       |                      MODIFIERS & ACCESS CONTROL                      |
